+++ CDLF/MaxPoolingLayer.cpp
@@ -5,7 +5,7 @@
 #include "MaxPoolingLayer.h"
 
 
-MaxPoolingLayer::MaxPoolingLayer(const int id, const string &name, const vector<int> &filterSize,
+MaxPoolingLayer::MaxPoolingLayer(const int id, const string &name, const vector<long> &filterSize,
                                  Layer *prevLayer, const int stride)
         : Layer(id, name, {}) {
 
@@ -57,35 +57,35 @@ void MaxPoolingLayer::forward() {
     const int N = m_filterSize.size();
     Tensor<float> &X = *m_prevLayer->m_pYTensor;
     if (2 == N) {
-        vector<int> Xc = {0, 0};
-        for (int i = 0; i < m_tensorSize[0]; ++i) {
+        vector<long> Xc = {0, 0};
+        for (long i = 0; i < m_tensorSize[0]; ++i) {
             Xc[0] += i * m_stride;
-            for (int j = 0; j < m_tensorSize[1]; ++j) {
+            for (long j = 0; j < m_tensorSize[1]; ++j) {
                 Xc[1] += j * m_stride;
                 m_pYTensor->e(i, j) = X.subTensorFromTopLeft(Xc, m_filterSize).max();
             }
         }
     } else if (3 == N) {
-        vector<int> Xc = {0, 0, 0};
-        for (int i = 0; i < m_tensorSize[0]; ++i) {
+        vector<long> Xc = {0, 0, 0};
+        for (long i = 0; i < m_tensorSize[0]; ++i) {
             Xc[0] += i * m_stride;
-            for (int j = 0; j < m_tensorSize[1]; ++j) {
+            for (long j = 0; j < m_tensorSize[1]; ++j) {
                 Xc[1] += j * m_stride;
-                for (int k = 0; k < m_tensorSize[2]; ++k) {
+                for (long k = 0; k < m_tensorSize[2]; ++k) {
                     Xc[2] += k * m_stride;
                     m_pYTensor->e(i, j, k) = X.subTensorFromTopLeft(Xc, m_filterSize).max();
                 }
             }
         }
     } else if (4 == N) {
-        vector<int> Xc = {0, 0, 0, 0};
-        for (int i = 0; i < m_tensorSize[0]; ++i) {
+        vector<long> Xc = {0, 0, 0, 0};
+        for (long i = 0; i < m_tensorSize[0]; ++i) {
             Xc[0] += i * m_stride;
-            for (int j = 0; j < m_tensorSize[1]; ++j) {
+            for (long j = 0; j < m_tensorSize[1]; ++j) {
                 Xc[1] += j * m_stride;
-                for (int k = 0; k < m_tensorSize[2]; ++k) {
+                for (long k = 0; k < m_tensorSize[2]; ++k) {
                     Xc[2] += k * m_stride;
-                    for (int l = 0; l < m_tensorSize[3]; ++l) {
+                    for (long l = 0; l < m_tensorSize[3]; ++l) {
                         Xc[3] += l * m_stride;
                         m_pYTensor->e(i, j, k, l) = X.subTensorFromTopLeft(Xc, m_filterSize).max();
                     }
@@ -105,49 +105,49 @@ void MaxPoolingLayer::backward() {
     Tensor<float> &dX = *(m_prevLayer->m_pdYTensor);
     Tensor<float> &X = *(m_prevLayer->m_pYTensor);
     if (2 == N) {
-        vector<int> Xc = {0, 0};
-        for (int i = 0; i < m_tensorSize[0]; ++i) {
+        vector<long> Xc = {0, 0};
+        for (long i = 0; i < m_tensorSize[0]; ++i) {
             Xc[0] += i * m_stride;
-            for (int j = 0; j < m_tensorSize[1]; ++j) {
+            for (long j = 0; j < m_tensorSize[1]; ++j) {
                 Xc[1] += j * m_stride;
                 const float maxValue = X.subTensorFromTopLeft(Xc, m_filterSize).max();
-                for (int ii = 0; ii < m_filterSize[0]; ++ii)
-                    for (int jj = 0; jj < m_filterSize[1]; ++jj)
+                for (long ii = 0; ii < m_filterSize[0]; ++ii)
+                    for (long jj = 0; jj < m_filterSize[1]; ++jj)
                         if (maxValue == X(Xc[0] + ii, Xc[1] + jj)) dX(Xc[0] + ii, Xc[1] + jj) += dLdy(i, j);
             }
         }
     } else if (3 == N) {
-        vector<int> Xc = {0, 0, 0};
-        for (int i = 0; i < m_tensorSize[0]; ++i) {
+        vector<long> Xc = {0, 0, 0};
+        for (long i = 0; i < m_tensorSize[0]; ++i) {
             Xc[0] += i * m_stride;
-            for (int j = 0; j < m_tensorSize[1]; ++j) {
+            for (long j = 0; j < m_tensorSize[1]; ++j) {
                 Xc[1] += j * m_stride;
-                for (int k = 0; k < m_tensorSize[2]; ++k) {
+                for (long k = 0; k < m_tensorSize[2]; ++k) {
                     Xc[2] += k * m_stride;
                     const float maxValue = X.subTensorFromTopLeft(Xc, m_filterSize).max();
-                    for (int ii = 0; ii < m_filterSize[0]; ++ii)
-                        for (int jj = 0; jj < m_filterSize[1]; ++jj)
-                            for (int kk = 0; kk < m_filterSize[2]; ++kk)
+                    for (long ii = 0; ii < m_filterSize[0]; ++ii)
+                        for (long jj = 0; jj < m_filterSize[1]; ++jj)
+                            for (long kk = 0; kk < m_filterSize[2]; ++kk)
                                 if (maxValue == X(Xc[0] + ii, Xc[1] + jj, Xc[2] + kk))
                                     dX(Xc[0] + ii, Xc[1] + jj, Xc[2] + kk) += dLdy(i, j, k);
                 }
             }
         }
     } else if (4 == N) {
-        vector<int> Xc = {0, 0, 0, 0};
-        for (int i = 0; i < m_tensorSize[0]; ++i) {
+        vector<long> Xc = {0, 0, 0, 0};
+        for (long i = 0; i < m_tensorSize[0]; ++i) {
             Xc[0] += i * m_stride;
-            for (int j = 0; j < m_tensorSize[1]; ++j) {
+            for (long j = 0; j < m_tensorSize[1]; ++j) {
                 Xc[1] += j * m_stride;
-                for (int k = 0; k < m_tensorSize[2]; ++k) {
+                for (long k = 0; k < m_tensorSize[2]; ++k) {
                     Xc[2] += k * m_stride;
-                    for (int l = 0; l < m_tensorSize[3]; ++l) {
+                    for (long l = 0; l < m_tensorSize[3]; ++l) {
                         Xc[3] += l * m_stride;
                         const float maxValue = X.subTensorFromTopLeft(Xc, m_filterSize).max();
-                        for (int ii = 0; ii < m_filterSize[0]; ++ii)
-                            for (int jj = 0; jj < m_filterSize[1]; ++jj)
-                                for (int kk = 0; kk < m_filterSize[2]; ++kk)
-                                    for (int ll = 0; ll < m_filterSize[3]; ++ll)
+                        for (long ii = 0; ii < m_filterSize[0]; ++ii)
+                            for (long jj = 0; jj < m_filterSize[1]; ++jj)
+                                for (long kk = 0; kk < m_filterSize[2]; ++kk)
+                                    for (long ll = 0; ll < m_filterSize[3]; ++ll)
                                         if (maxValue == X(Xc[0] + ii, Xc[1] + jj, Xc[2] + kk, Xc[3] + ll))
                                             dX(Xc[0] + ii, Xc[1] + jj, Xc[2] + kk, Xc[3] + ll) += dLdy(i, j, k, l);
                     }
